chore(skeleton): split demo
