+++ packages/nutui-taro-demo/src/exhibition/pages/skeleton/basic.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-skeleton width="200px" height="15px" animated></nut-skeleton>
+</template>
