+++ packages/nutui-taro-demo/src/exhibition/pages/skeleton/row.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-skeleton width="250px" height="15px" animated row="3"> </nut-skeleton>
+</template>
