+++ src/packages/__VUE/skeleton/demo/index.vue
@@ -21,7 +21,7 @@
 </template>
 
 <script setup lang="ts">
-import { useTranslate } from '../../../utils';
+import { useTranslate } from '@/sites/utils';
 import Basic from './basic.vue';
 import Row from './row.vue';
 import Avatar from './avatar.vue';
