+++ src/packages/__VUE/skeleton/demo/picture.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="skeleton-picture">
+    <nut-skeleton width="250px" height="15px" title animated row="3" class="item"> </nut-skeleton>
+    <nut-skeleton width="250px" height="15px" title animated row="3" class="item"> </nut-skeleton>
+  </div>
+</template>
+<style>
+.skeleton-picture {
+  display: flex;
+  justify-content: space-between;
+  .item {
+    width: 47%;
+  }
+}
+</style>
