+++ src/packages/__VUE/skeleton/demo/round.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-skeleton width="250px" height="15px" animated round></nut-skeleton>
+</template>
