+++ src/packages/__VUE/skeleton/demo/row.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-skeleton width="250px" height="15px" animated row="3"> </nut-skeleton>
+</template>
