+++ src/packages/__VUE/skeleton/doc.en-US.md
@@ -16,134 +16,27 @@ app.use(Skeleton);
 
 ### Basic usage
 
-:::demo
+> demo: skeleton basic
 
-```vue
-<template>
-  <nut-skeleton width="250px" height="15px" animated> </nut-skeleton>
-</template>
-```
+### Multi-lines
 
-:::
+> demo: skeleton row
 
-### Incoming multiline
+### Show Avatar
 
-:::demo
+> demo: skeleton avatar
 
-```vue
-<template>
-  <nut-skeleton width="250px" height="15px" title animated row="3"> </nut-skeleton>
-</template>
-```
+### Round Style
 
-:::
-
-### Show Faces
-
-:::demo
-
-```vue
-<template>
-  <nut-skeleton width="250px" height="15px" title animated avatar avatarSize="60px" row="3"> </nut-skeleton>
-</template>
-```
-
-:::
-
-### Title paragraph fillet style
-
-:::demo
-
-```vue
-<template>
-  <nut-skeleton width="250px" height="15px" animated round></nut-skeleton>
-</template>
-```
-
-:::
+> demo: skeleton round
 
 ### Picture combination
 
-:::demo
-
-```vue
-<template>
-  <div class="pic-compose">
-    <nut-skeleton width="250px" height="15px" title animated row="3" class="item"> </nut-skeleton>
-    <nut-skeleton width="250px" height="15px" title animated row="3" class="item"> </nut-skeleton>
-  </div>
-</template>
-<style>
-.pic-compose {
-  display: flex;
-  justify-content: space-between;
-}
-.item {
-  width: 47%;
-}
-</style>
-```
-
-:::
+> demo: skeleton picture
 
 ### Display subcomponents
 
-:::demo
-
-```vue
-<template>
-  <div class="content">
-    <nut-switch v-model="checked" />
-    <nut-skeleton width="250px" height="15px" title animated avatar row="3" :loading="!checked">
-      <div class="container">
-        <nut-avatar size="50">
-          <img
-            src="https://img14.360buyimg.com/imagetools/jfs/t1/167902/2/8762/791358/603742d7E9b4275e3/e09d8f9a8bf4c0ef.png"
-          />
-        </nut-avatar>
-      <div class="right-content">
-        <div class="title">NutUI</div>
-        <div class="desc"
-            >A set of JD style lightweight mobile terminal Vue group library provides rich basic components and business components to help developers quickly build mobile applications.
-        </div>
-      </div>
-    </nut-skeleton>
-  </div>
-</template>
-
-<script setup>
-import { ref } from 'vue';
-const checked = ref(false);
-</script>
-
-<style>
-  .nut-switch {
-    display: flex;
-    margin: 0 16px 8px 0;
-  }
-  .container {
-    display: flex;
-  }
-  .right-content {
-    margin-left: 19px;
-    font-family: PingFangSC;
-    display: flex;
-    flex-direction: column;
-  }
-  .title {
-    font-size: 14px;
-    color: rgba(51, 51, 51, 1);
-  }
-  .desc {
-    margin-top: 10px;
-    font-size: 13px;
-    color: rgba(154, 155, 157, 1);
-  }
-</style>
-
-```
-
-:::
+> demo: skeleton sub
 
 ## API
