+++ src/packages/__VUE/skeleton/doc.taro.md
@@ -16,137 +16,27 @@ app.use(Skeleton);
 
 ### 基础用法
 
-:::demo
-
-```vue
-<template>
-  <nut-skeleton width="250px" height="15px" animated> </nut-skeleton>
-</template>
-```
-
-:::
+> demo: skeleton basic exhibition
 
 ### 传入多行
 
-:::demo
-
-```vue
-<template>
-  <nut-skeleton width="250px" height="15px" title animated row="3"> </nut-skeleton>
-</template>
-```
-
-:::
+> demo: skeleton row exhibition
 
 ### 显示头像
 
-:::demo
-
-```vue
-<template>
-  <nut-skeleton width="250px" height="15px" title animated avatar avatarSize="60px" row="3"> </nut-skeleton>
-</template>
-```
-
-:::
+> demo: skeleton avatar exhibition
 
 ### 标题段落圆角风格
 
-:::demo
-
-```vue
-<template>
-  <nut-skeleton width="250px" height="15px" animated round></nut-skeleton>
-</template>
-```
-
-:::
+> demo: skeleton round exhibition
 
 ### 图片组合
 
-:::demo
-
-```vue
-<template>
-  <div class="pic-compose">
-    <nut-skeleton width="250px" height="15px" title animated row="3" class="item"> </nut-skeleton>
-    <nut-skeleton width="250px" height="15px" title animated row="3" class="item"> </nut-skeleton>
-  </div>
-</template>
-```
-
-```css
-<style>
-.pic-compose {
-  display: flex;
-  justify-content: space-between;
-  .item {
-    width: 47%;
-  }
-}
-</style>
-```
-
-:::
+> demo: skeleton picture exhibition
 
 ### 显示子组件
 
-:::demo
-
-```vue
-<template>
-  <div class="content">
-    <nut-switch v-model="checked" />
-    <nut-skeleton width="250px" height="15px" title animated avatar row="3" :loading="!checked">
-      <div class="container">
-        <nut-avatar size="50">
-          <img
-            src="https://img14.360buyimg.com/imagetools/jfs/t1/167902/2/8762/791358/603742d7E9b4275e3/e09d8f9a8bf4c0ef.png"
-          />
-        </nut-avatar>
-        <div class="right-content">
-          <div class="title">NutUI</div>
-          <div class="desc"
-            >一套京东风格的轻量级移动端Vue组库，提供丰富的基础组件和业务组件，帮助开发者快速搭建移动应用。</div
-          >
-        </div>
-      </div>
-    </nut-skeleton>
-  </div>
-</template>
-
-<script setup>
-import { ref } from 'vue';
-const checked = ref(false);
-</script>
-
-<style>
-.nut-switch {
-  display: flex;
-  margin: 0 16px 8px 0;
-}
-.container {
-  display: flex;
-}
-.right-content {
-  margin-left: 19px;
-  font-family: PingFangSC;
-  display: flex;
-  flex-direction: column;
-}
-.title {
-  font-size: 14px;
-  color: rgba(51, 51, 51, 1);
-}
-.desc {
-  margin-top: 10px;
-  font-size: 13px;
-  color: rgba(154, 155, 157, 1);
-}
-</style>
-```
-
-:::
+> demo: skeleton sub exhibition
 
 ## API
 
@@ -162,8 +52,8 @@ const checked = ref(false);
 | avatar-shape | 头像形状：正方形/圆形 | string | `round` |
 | avatar-size | 头像大小 | string | `50px` |
 | round | 标题/段落是否采用圆角风格 | boolean | `false` |
-| row | 设置段落行数 | string | `1` |
-| title | 是否显示段落标题 | boolean | `true` |
+| title | 是否显示段落标题（第一行显示） | boolean | `true` |
+| row | 设置段落行数（title 下方展示） | string | `1` |
 
 ### Slots
